Make hyperlane compatible with multiaddress
